+++ src/worker/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "boot")]
+mod boot;
 mod local_file;
 mod memory;
 #[cfg(feature = "postgres")]
@@ -6,10 +8,12 @@ mod queue;
 mod runner;
 mod task;
 
+#[cfg(feature = "boot")]
+pub use boot::BootFlowTaskManager;
 pub use local_file::{LocalFileDeadLetteredTask, LocalFileFlowTaskQueue};
 pub use memory::InMemoryFlowTaskQueue;
 #[cfg(feature = "postgres")]
 pub use postgres::{PostgresDeadLetteredTask, PostgresFlowTaskQueue};
-pub use queue::FlowTaskQueue;
+pub use queue::{FlowTaskDispatcher, FlowTaskQueue};
 pub use runner::FlowWorker;
 pub use task::{FlowTask, FlowTaskLease, FlowTaskOutcome};